Move routesUtils to src/utilities/
