+++ app/packs/src/components/elements/samples/measurementsTab/MeasurementsList.js
@@ -5,7 +5,7 @@ import { Button, Glyphicon } from 'react-bootstrap';
 import { observer } from 'mobx-react';
 
 import { StoreContext } from 'src/stores/mobx/RootStore';
-import { researchPlanShowOrNew } from 'src/components/routesUtils';
+import { researchPlanShowOrNew } from 'src/utilities/routesUtils';
 import ConfirmDeletionModal from 'src/components/elements/samples/measurementsTab/ConfirmDeletionModal';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
