+++ app/packs/src/components/elements/samples/measurementsTab/MeasurementsTable.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { observer } from 'mobx-react';
 
 import { StoreContext } from 'src/stores/mobx/RootStore';
-import { researchPlanShowOrNew } from 'src/components/routesUtils';
+import { researchPlanShowOrNew } from 'src/utilities/routesUtils';
 
 class MeasurementsTable extends Component {
   static propTypes = {
