+++ app/packs/src/components/elements/screens/ScreenWellplates.js
@@ -5,7 +5,7 @@ import { DropTarget } from 'react-dnd';
 import Aviator from 'aviator';
 import DragDropItemTypes from 'src/components/DragDropItemTypes';
 import UIStore from 'src/stores/alt/stores/UIStore';
-import { wellplateShowOrNew } from 'src/components/routesUtils';
+import { wellplateShowOrNew } from 'src/utilities/routesUtils';
 import QuillViewer from 'src/components/QuillViewer';
 
 const target = {
