+++ app/packs/src/components/elements/screens/researchPlansTab/EmbeddedResearchPlanDetails.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Panel, ListGroup, ListGroupItem, Button, ButtonGroup, Tooltip, Overlay, OverlayTrigger, Dropdown, MenuItem } from 'react-bootstrap';
 import Aviator from 'aviator';
 import UIStore from 'src/stores/alt/stores/UIStore';
-import { researchPlanShowOrNew } from 'src/components/routesUtils';
+import { researchPlanShowOrNew } from 'src/utilities/routesUtils';
 import ElementCollectionLabels from 'src/components/elementsList/ElementCollectionLabels';
 import ResearchPlansFetcher from 'src/fetchers/ResearchPlansFetcher';
 import ResearchPlan from 'src/models/ResearchPlan';
